Flipped camera fix & renamed OSC variables
Removed an adjustment to fix a (suddenly) flipped axis on the camera when using the vive tracker (up is down and vice versa) that wasn't properly documented and lead to confusion. Will integrate a function to allow you to adjust for this with a single button press, but I'm gonna test this first. If you encounter a problem with flipped axis, please contact us and I will guide you through how to do it.

Also I renamed serverIP to computerIP and controlIP to phoneIP and added some tooltips to make it clear where to enter what.